Allow double for battery residual

+++ app/src/main/java/com/alpriest/energystats/models/BatteryResponse.kt
@@ -3,7 +3,7 @@ package com.alpriest.energystats.models
 data class BatteryResponse(
     val power: Double,
     val soc: Int,
-    val residual: Int,
+    val residual: Double,
     val temperature: Double
 )
